+++ packages/api-docs/config/api-docs.js
@@ -7,6 +7,7 @@ const apiDocs = require('../lib/index');
 
 /** @type {apiDocs.IPageJsonOptions} */
 module.exports = {
+  // NOTE: when adding new package to this list, also add package dep in package.json.
   apiJsonPaths: [
     path.resolve(__dirname, '../../styling/dist/styling.api.json'),
     path.resolve(__dirname, '../../theme/dist/theme.api.json'),
@@ -19,6 +20,7 @@ module.exports = {
     path.resolve(__dirname, '../../react-link/dist/react-link.api.json'),
     path.resolve(__dirname, '../../react-slider/dist/react-slider.api.json'),
     path.resolve(__dirname, '../../react-tabs/dist/react-tabs.api.json'),
+    path.resolve(__dirname, '../../react-theme-provider/dist/react-theme-provider.api.json'),
     path.resolve(__dirname, '../../react-toggle/dist/react-toggle.api.json'),
     path.resolve(__dirname, '../../react/dist/react.api.json'),
     path.resolve(__dirname, '../../date-time-utilities/dist/date-time-utilities.api.json'),