+++ packages/fluentui/docs/src/examples/components/Tree/Usage/index.tsx
@@ -25,6 +25,11 @@ const Usage = () => (
       description="A multiselect Tree."
       examplePath="components/Tree/Usage/TreeMultiselectExample"
     />
+    <ComponentExample
+      title="Multi select As a List"
+      description="A multiselect Tree with list accessibility roles."
+      examplePath="components/Tree/Usage/TreeMultiselectAsListExample"
+    />
   </ExampleSection>
 );