+++ packages/react-internal/src/components/ContextualMenu/index.ts
@@ -4,4 +4,3 @@ export * from './ContextualMenu.types';
 export * from './ContextualMenuItem';
 export * from './ContextualMenuItem.base';
 export * from './ContextualMenuItem.types';
-export { getMenuItemStyles } from './ContextualMenu.cnstyles';