+++ packages/react-internal/src/components/FloatingPicker/PeoplePicker/FloatingPeoplePicker.tsx
@@ -5,6 +5,7 @@ import { SuggestionItemNormal } from './PeoplePickerItems/SuggestionItemDefault'
 import { IPersonaProps } from '../../../Persona';
 import './PeoplePicker.scss';
 import { IBasePickerSuggestionsProps, ISuggestionModel } from '../../../Pickers';
+/* eslint-disable */
 
 /**
  * {@docCategory FloatingPeoplePicker}