+++ packages/react-internal/src/components/Layer/index.ts
@@ -2,3 +2,4 @@ export * from './Layer';
 export * from './Layer.base';
 export * from './Layer.types';
 export * from './LayerHost';
+export * from './LayerHost.types';