+++ packages/react-next/src/components/FloatingPicker/index.ts
@@ -1,7 +1,0 @@
-export * from './BaseFloatingPicker';
-export * from './BaseFloatingPicker.types';
-export * from './PeoplePicker/FloatingPeoplePicker';
-export * from './Suggestions/SuggestionsStore';
-export * from './Suggestions/SuggestionsControl';
-export * from './Suggestions/SuggestionsCore';
-export * from './Suggestions/Suggestions.types';