+++ scripts/tasks/jest.js
@@ -3,10 +3,17 @@
 const { jestTask, argv } = require('just-scripts');
 const path = require('path');
 
+const commonArgs = () => {
+  const args = argv();
+  return {
+    ...((process.env.TF_BUILD || process.env.LAGE_PACKAGE_NAME) && { runInBand: true }),
+    ...(args.u || args.updateSnapshot ? { updateSnapshot: true } : undefined),
+  };
+};
+
 exports.jest = () =>
   jestTask({
-    ...((process.env.TF_BUILD || process.env.LAGE_PACKAGE) && { runInBand: true }),
-    ...(argv().u || argv().updateSnapshot ? { updateSnapshot: true } : undefined),
+    ...commonArgs(),
     env: {
       ...process.env,
       NODE_ENV: 'test',
@@ -22,8 +29,7 @@ exports.jestDom = () =>
 exports.jestWatch = () => {
   const args = argv();
   return jestTask({
-    ...((process.env.TF_BUILD || process.env.LAGE_PACKAGE) && { runInBand: true }),
-    ...(args.u || args.updateSnapshot ? { updateSnapshot: true } : undefined),
+    ...commonArgs(),
     watch: true,
     _: ['-i', ...(args._ || []).filter(arg => arg !== 'jest-watch')],
   });